move SpreedlyErrors to domain model package
